readd supp for winpe
re-add support for winpe.

+++ ExternalWIN/GetWIMTarg.vbs
@@ -1,5 +1,6 @@
 wim = WScript.Arguments(0)
 index = WScript.Arguments(1)
+winpe = UCase(WScript.Arguments(2))
 Set objShell = CreateObject("WScript.Shell")
 Set objExec = objShell.Exec("dism /List-Image /ImageFile:""" & wim & """ /index:" & index)
 pid = objExec.ProcessID
@@ -8,12 +9,20 @@ Do While Not objExec.StdOut.AtEndOfStream
 	' IF the WIM Image Doesn't Contain The Target then Assume it's the entire C Drive
 	If Right(strLine, 1) = "\" And Len(strLine) > 2 Then
 		WScript.Echo "\"
-		objShell.Run "taskkill /F /PID " & pid, 0, True
+		IF winpe = "TRUE" Then
+			objExec.Terminate
+		Else
+			objShell.Run "taskkill /F /PID " & pid, 0, True
+		End If
 		WScript.Quit
 	End If
     If InStr(1, strLine, "\EXTWNCAP$") = 1 Then
 		WScript.Echo strLine
-		objShell.Run "taskkill /F /PID " & pid, 0, True
+		IF winpe = "TRUE" Then
+			objExec.Terminate
+		Else
+			objShell.Run "taskkill /F /PID " & pid, 0, True
+		End If
 		WScript.Quit
     End If
 Loop
